+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 THe format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.7]() (2024-06-05)
+
+### Added
+
+- Added config sidebar to change CV colors, shape of profile picture, and weight of headlines.
+- Added description to the project, including references to the original project and course.
+
 ## [0.0.6]() (2024-06-05)
 
 ### Added